move css files to separate folder (styles) and modified import statements
