+++ frontend/src/App.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import './App.css'
+import '../styles/App.css'
 import LocationSelector from '../components/LocationSelector';
 
 const App = () => {
