+++ frontend/src/main.jsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
-import './index.css'
+import '../styles/index.css'
 import Login from './Login'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    {/* <Login /> */}
+    <Login />
     <App />
   </React.StrictMode>
 )
